# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invoice Tracker API",
    "description": "REST API for the invoice-tracker application. Hand-constructed by the documentation agent for FEAT-20260513-03 because the backend cannot start without a running Postgres instance. Regenerate once `docker-compose up` is available: `curl http://localhost:8080/v3/api-docs | python3 -m json.tool > docs/openapi.json`.",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    { "basicAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/invoices": {
      "post": {
        "tags": ["Invoices"],
        "summary": "Create a new invoice",
        "operationId": "createInvoice",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateInvoiceRequest" },
              "example": {
                "number": "INV-2026-001",
                "clientId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "issueDate": "2026-05-14",
                "dueDate": "2026-06-13",
                "taxRate": 0.20,
                "lines": [
                  { "description": "Web development", "quantity": 10, "unitPrice": 150.00 }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created",
            "headers": {
              "Location": {
                "description": "URL of the created invoice",
                "schema": { "type": "string", "format": "uri" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvoiceResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "get": {
        "tags": ["Invoices"],
        "summary": "List invoices",
        "operationId": "listInvoices",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "description": "Filter by client UUID",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt,desc" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated invoice list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvoicePage" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/invoices/{id}": {
      "get": {
        "tags": ["Invoices"],
        "summary": "Get an invoice by ID",
        "operationId": "getInvoice",
        "parameters": [
          { "$ref": "#/components/parameters/InvoiceId" }
        ],
        "responses": {
          "200": {
            "description": "Invoice found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvoiceResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/InvoiceNotFound" }
        }
      }
    },
    "/api/v1/invoices/{id}/pdf": {
      "get": {
        "tags": ["Invoices"],
        "summary": "Download invoice as PDF",
        "operationId": "getInvoicePdf",
        "parameters": [
          { "$ref": "#/components/parameters/InvoiceId" }
        ],
        "responses": {
          "200": {
            "description": "PDF document stream",
            "headers": {
              "Content-Disposition": {
                "schema": { "type": "string" },
                "example": "inline; filename=\"invoice-INV-2026-001.pdf\""
              },
              "Cache-Control": {
                "schema": { "type": "string" },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/InvoiceNotFound" },
          "502": { "$ref": "#/components/responses/PdfConversionFailed" }
        }
      }
    },
    "/api/v1/invoices/{id}/send-email": {
      "post": {
        "tags": ["Invoices"],
        "summary": "Send invoice by email to client",
        "operationId": "sendInvoiceEmail",
        "parameters": [
          { "$ref": "#/components/parameters/InvoiceId" }
        ],
        "responses": {
          "200": {
            "description": "Email sent — returns lastSentAt timestamp",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SendEmailResponse" },
                "example": { "lastSentAt": "2026-05-14T10:45:00Z" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/InvoiceNotFound" },
          "422": { "$ref": "#/components/responses/InvoiceHasNoRecipient" },
          "502": { "$ref": "#/components/responses/EmailDeliveryFailed" }
        }
      }
    },
...
    "/api/v1/settings/invoice-template": {
      "post": {
        "tags": ["Settings"],
        "summary": "Upload a new DOCX invoice template",
        "operationId": "uploadInvoiceTemplate",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": ".docx file, max 5 MB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadTemplateResponse" },
                "example": {
                  "filename": "invoice-template.docx",
                  "size": 12345,
                  "uploadedAt": "2026-05-14T10:00:00Z"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/TemplateTooLarge" },
          "415": { "$ref": "#/components/responses/InvalidTemplateType" }
        }
      }
    },
    "/api/v1/settings/invoice-template/preview": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get active template metadata",
        "operationId": "getInvoiceTemplatePreview",
        "responses": {
          "200": {
            "description": "Template metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TemplateMetadataResponse" },
                "example": {
                  "filename": "invoice-template.docx",
                  "size": 12345,
                  "uploadedAt": "2026-05-14T10:00:00Z",
                  "isDefault": false
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/settings/invoice-template/download": {
      "get": {
        "tags": ["Settings"],
        "summary": "Download the active invoice template",
        "operationId": "downloadInvoiceTemplate",
        "responses": {
          "200": {
            "description": "Active template DOCX file",
            "headers": {
              "Content-Disposition": {
                "schema": { "type": "string" },
                "example": "attachment; filename=\"invoice-template.docx\""
              },
              "Cache-Control": {
                "schema": { "type": "string" },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/TemplateNotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
...
      "InvoiceId": {
        "name": "id",
        "in": "path",
        "description": "UUID of the invoice",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "schemas": {
...
      "CreateInvoiceRequest": {
        "type": "object",
        "required": ["number", "clientId", "issueDate", "dueDate", "lines"],
        "properties": {
          "number": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "Invoice number (e.g. INV-2026-001)"
          },
          "clientId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the associated client"
          },
          "issueDate": {
            "type": "string",
            "format": "date",
            "description": "Issue date (ISO-8601 date)"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "Payment due date (ISO-8601 date)"
          },
          "taxRate": {
            "type": "number",
            "format": "decimal",
            "description": "Tax rate as a decimal fraction (e.g. 0.20 for 20%)",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "items": { "$ref": "#/components/schemas/CreateInvoiceLineRequest" }
          }
        }
      },
      "CreateInvoiceLineRequest": {
        "type": "object",
        "required": ["description", "quantity", "unitPrice"],
        "properties": {
          "description": { "type": "string", "minLength": 1 },
          "quantity": { "type": "number", "minimum": 0 },
          "unitPrice": { "type": "number", "minimum": 0 }
        }
      },
      "InvoiceLineDto": {
        "type": "object",
        "required": ["id", "description", "quantity", "unitPrice", "lineTotal"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "description": { "type": "string" },
          "quantity": { "type": "number" },
          "unitPrice": { "type": "number" },
          "lineTotal": { "type": "number" }
        }
      },
      "InvoiceResponse": {
        "type": "object",
        "required": ["id", "number", "clientId", "issueDate", "dueDate", "lines", "subtotal", "total", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "number": { "type": "string" },
          "clientId": { "type": "string", "format": "uuid" },
          "clientEmail": { "type": "string", "format": "email", "nullable": true },
          "issueDate": { "type": "string", "format": "date" },
          "dueDate": { "type": "string", "format": "date" },
          "lines": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/InvoiceLineDto" }
          },
          "taxRate": { "type": "number", "nullable": true },
          "subtotal": { "type": "number" },
          "total": { "type": "number" },
          "lastSentAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "InvoicePage": {
        "type": "object",
        "required": ["content", "page", "size", "totalElements", "totalPages"],
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/InvoiceResponse" }
          },
          "page": { "type": "integer", "minimum": 0 },
          "size": { "type": "integer", "minimum": 1 },
          "totalElements": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
      "SendEmailResponse": {
        "type": "object",
        "required": ["lastSentAt"],
        "properties": {
          "lastSentAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the email was successfully sent"
          }
        }
      },
      "UploadTemplateResponse": {
        "type": "object",
        "required": ["filename", "size", "uploadedAt"],
        "properties": {
          "filename": { "type": "string", "example": "invoice-template.docx" },
          "size": { "type": "integer", "description": "File size in bytes" },
          "uploadedAt": { "type": "string", "format": "date-time" }
        }
      },
      "TemplateMetadataResponse": {
        "type": "object",
        "required": ["filename", "size", "uploadedAt", "isDefault"],
        "properties": {
          "filename": { "type": "string", "example": "invoice-template.docx" },
          "size": { "type": "integer", "description": "File size in bytes" },
          "uploadedAt": { "type": "string", "format": "date-time" },
          "isDefault": {
            "type": "boolean",
            "description": "true when serving the bundled classpath template; false when a user-uploaded template is active"
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 7807 Problem Detail",
        "properties": {
          "type": { "type": "string", "example": "about:blank" },
          "title": { "type": "string", "example": "Conflict" },
          "status": { "type": "integer", "example": 409 },
          "detail": { "type": "string", "example": "A client with this email already exists." },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "MISSING_REQUEST_PART",
              "CLIENT_NOT_FOUND",
              "CLIENT_EMAIL_TAKEN",
              "USER_EMAIL_TAKEN",
              "INVOICE_NOT_FOUND",
              "INVOICE_HAS_NO_RECIPIENT",
              "EMAIL_DELIVERY_FAILED",
              "PDF_CONVERSION_FAILED",
              "PDF_CONVERSION_BUSY",
              "INVALID_TEMPLATE_TYPE",
              "TEMPLATE_TOO_LARGE",
              "TEMPLATE_NOT_FOUND",
              "UNAUTHENTICATED",
              "INTERNAL_ERROR"
            ]
          },
          "instance": { "type": "string", "example": "/api/v1/clients" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid HTTP Basic credentials",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Unauthorized", "status": 401, "code": "UNAUTHENTICATED" }
          }
        }
      },
      "ValidationFailed": {
        "description": "Bean-validation failure — one or more fields invalid",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Bad Request", "status": 400, "code": "VALIDATION_FAILED" }
          }
        }
      },
...
      "InvoiceNotFound": {
        "description": "No invoice with the given UUID",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Not Found", "status": 404, "code": "INVOICE_NOT_FOUND" }
          }
        }
      },
      "InvoiceHasNoRecipient": {
        "description": "The invoice's client has no email address",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Unprocessable Entity", "status": 422, "code": "INVOICE_HAS_NO_RECIPIENT" }
          }
        }
      },
      "EmailDeliveryFailed": {
        "description": "SMTP relay rejected the message or timed out",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Bad Gateway", "status": 502, "code": "EMAIL_DELIVERY_FAILED" }
          }
        }
      },
      "PdfConversionFailed": {
        "description": "LibreOffice headless exited non-zero or timed out (20 s)",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Bad Gateway", "status": 502, "code": "PDF_CONVERSION_FAILED" }
          }
        }
      },
...
      "TemplateTooLarge": {
        "description": "Uploaded template exceeds the 5 MB limit",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Payload Too Large", "status": 413, "code": "TEMPLATE_TOO_LARGE" }
          }
        }
      },
      "InvalidTemplateType": {
        "description": "Uploaded file is not a valid .docx (wrong magic bytes, missing word/document.xml, or external relationship references detected)",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Unsupported Media Type", "status": 415, "code": "INVALID_TEMPLATE_TYPE" }
          }
        }
      },
      "TemplateNotFound": {
        "description": "No template on the filesystem and classpath default is also missing (should not occur in normal operation)",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": { "type": "about:blank", "title": "Not Found", "status": 404, "code": "TEMPLATE_NOT_FOUND" }
          }
        }
      }
    }
  }
}